test: moved ssl tests into existing suite
